+++ CHANGELOG.md
@@ -8,6 +8,10 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+### Fixed
+
+- Compute `origin_url` from consumer URL instead of `HTTP_REFERER`
+
 ## [1.3.0] - 2024-04-25
 
 ### Added